requirements.txt
beautifulsoup4==4.13.4
httpx==0.28.1
pydantic==2.11.1
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
typer==0.15.2

